+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 1.0.0
 
-* Check ReadMe page
+* Check README